refactor: markets
