+++ apps/lostark-hands-next/app/markets/page.tsx
@@ -2,14 +2,14 @@ import ServerWrapper from '@/app/server-wrapper';
 
 import { getOptionsApi } from '@/service/markets';
 
-import Markets from '@/app/markets/@component';
+import Search from '@/app/markets/@component/search';
 
 const Page = () => {
 	return (
 		<div className="space-y-[16px] px-[16px] pb-[16px]">
 			<ServerWrapper
 				apiPromise={getOptionsApi()}
-				render={Markets}
+				render={Search}
 			/>
 		</div>
 	);
